requirements.txt
aiokafka==0.10.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
bcrypt==5.0.0
bidict==0.23.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.110.0
h11==0.16.0
httptools==0.7.1
idna==3.11
mysql-connector-python==8.3.0
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-engineio==4.12.3
python-jose==3.3.0
python-multipart==0.0.21
python-socketio==5.11.2
PyYAML==6.0.3
redis==5.0.1
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.27.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
wsproto==1.3.2
